+++ app/policies/guild_invite_policy.rb
@@ -2,6 +2,10 @@
 class GuildInvitePolicy < ApplicationPolicy
   authorize :guild, :character
 
+  def index?
+    new?
+  end
+
   def new?
     return user.any_role?(guild.id, 'gm', 'rl') if record == 'true'
     character.user_id == user.id && character.guild_id.nil?