+++ frontend/components/characters/characters.js
@@ -1,0 +1,1 @@
+import "./characters.scss";