+++ frontend/packs/invite_form_for_character.jsx
@@ -1,0 +1,4 @@
+import InviteFormForCharacter from "components_react/invite_form_for_character/invite_form_for_character";
+import WebpackerReact from "webpacker-react";
+
+WebpackerReact.setup({InviteFormForCharacter});